+++ src/Marvin/Adapter/Slack/Internal/Types.hs
@@ -22,7 +22,8 @@ import           Util
 
 
 jsonParseURI :: Value -> Parser URI
-jsonParseURI =  withText "expected text" $ maybe (fail "string not parseable as uri") return . parseURI . T.unpack
+jsonParseURI = withText "expected text"
+    $ maybe (fail "string not parseable as uri") return . parseURI . T.unpack
 
 
 data RTMData = RTMData
@@ -34,9 +35,11 @@ type APIResponse a = Either String a
 
 
 -- | Identifier for a user (internal and not equal to the username)
-newtype SlackUserId = SlackUserId { unwrapSlackUserId :: T.Text } deriving (IsString, Eq, Hashable)
+newtype SlackUserId = SlackUserId { unwrapSlackUserId :: T.Text }
+    deriving (IsString, Eq, Hashable)
 -- | Identifier for a channel (internal and not equal to the channel name)
-newtype SlackChannelId = SlackChannelId { unwrapSlackChannelId :: T.Text } deriving (IsString, Eq, Show, Hashable)
+newtype SlackChannelId = SlackChannelId { unwrapSlackChannelId :: T.Text }
+    deriving (IsString, Eq, Show, Hashable)
 
 
 deriveJSON defaultOptions { unwrapUnaryRecords = True } ''SlackUserId
@@ -47,11 +50,21 @@ class HasIdValue s a | s -> a where idValue :: Lens' s a
 class HasNameResolver s a | s -> a where nameResolver :: Lens' s a
 class HasInfoCache s a | s -> a where infoCache :: Lens' s a
 class HasCreated s a | s -> a where created :: Lens' s a
+class HasChanType s a | s -> a where chanType :: Lens' s a
+-- class HasMemberCount s a | s -> a where memberCount :: Lens' s a
+
+data ChannelType
+    = PublicChannel
+    | PrivateChannel
+    | IMChannel
+    deriving Show
 
 data LimitedChannelInfo = LimitedChannelInfo
-    { limitedChannelInfoIdValue :: SlackChannelId
-    , limitedChannelInfoName    :: L.Text
-    , limitedChannelInfoTopic   :: L.Text
+    { limitedChannelInfoIdValue     :: SlackChannelId
+    , limitedChannelInfoName        :: Maybe L.Text
+    , limitedChannelInfoTopic       :: Maybe L.Text
+    -- , limitedChannelInfoChanType    :: ChannelType
+    -- , limitedChannelInfoMemberCount :: Int
     } deriving Show
 
 data UserInfo = UserInfo
@@ -72,7 +85,7 @@ data UserCache = UserCache
     , userCacheNameResolver :: HashMap L.Text UserInfo
     }
 
--- | Adapter for interacting with Slack API\'s. Polymorphic over the method for retrieving events.
+-- | Adapter for interacting with Slack API's. Polymorphic over the method for retrieving events.
 data SlackAdapter a = SlackAdapter
     { slackAdapterChannelCache  :: MVar ChannelCache
     , slackAdapterUserInfoCache :: MVar UserCache
@@ -188,9 +201,24 @@ lciParser :: Value -> Parser LimitedChannelInfo
 lciParser = withObject "expected object" $ \o ->
     LimitedChannelInfo
         <$> o .: "id"
-        <*> o .: "name"
-        <*> (o .: "topic" >>= withObject "object" (.: "value"))
+        <*> o .:? "name"
+        <*> (o .:? "topic" >>= maybe (return Nothing) (fmap Just . withObject "object" (.: "value")))
+        -- <*> parseType o
+        -- <*> o .: "num_members"
+  -- where
+  --   parseType o =
+  --       ifM (o .: "is_channel")
+  --           (pure PublicChannel)
+  --           $ ifM (o .: "is_group")
+  --               (pure PrivateChannel)
+  --               $ ifM (o .: "is_im")
+  --                   (pure IMChannel)
+  --                   (fail "Expected one of: [is_channel, is_group, is_im] to be true.")
+
+
+ifM :: Monad m => m Bool -> m a -> m a -> m a
+ifM cond then_ else_ = cond >>= \c -> if c then then_ else else_
+
 
 lciListParser :: Value -> Parser [LimitedChannelInfo]
 lciListParser = withArray "array" $ fmap toList . mapM lciParser
-