+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.4 - 23.09.2017
+
+- Fixed an old bug when getting information about unnamed channels
+
 # 0.2.3 - 05.03.2017
 
 - A big thanks to @lubomir for testing the IRC adapter and contributing the following changes